do not remove existing README file (#22)
The Debian package ships a useful README file and it doesn't seem
relevant to forcibly remove it.

This will keep the existing file if it already exists, otherwise it will just create an empty
file. That still seems preferable to removing the file altogether or
trying to synchronise the file's content with the module eternally.